+++ src/entities/core-entities/enums/player-type.ts
@@ -1,0 +1,10 @@
+export enum PlayerType {
+  Player = 1,
+  Other = 2,
+  Coach = 3,
+}
+
+
+
+
+