+++ src/entities/core-entities/fixture/league.ts
@@ -1,4 +1,6 @@
-import { Expose } from 'class-transformer';
+import { Expose, Type } from 'class-transformer';
+
+import { IdNNameRecord } from '@lsports/entities/common';
 
 export class League {
   @Expose({ name: 'Id' })
@@ -6,4 +8,24 @@ export class League {
 
   @Expose({ name: 'Name' })
   public name?: string;
+
+  @Expose({ name: 'Tour' })
+  @Type(() => IdNNameRecord)
+  public tour?: IdNNameRecord;
+
+  @Expose({ name: 'AgeCategory' })
+  public ageCategory?: number;
+
+  @Expose({ name: 'Gender' })
+  public gender?: number;
+
+  @Expose({ name: 'Type' })
+  public type?: number;
+
+  @Expose({ name: 'NumberOfPeriods' })
+  public numberOfPeriods?: number;
+
+  @Expose({ name: 'SportCategory' })
+  @Type(() => IdNNameRecord)
+  public sportCategory?: IdNNameRecord;
 }